A grouped data table.

The easiest way to create a grouped data table is to call the group_by method on a data table or tbl: this will take care of capturing the unevalated expressions for you.

grouped_dt(data, vars, copy = TRUE)

is.grouped_dt(x)

Arguments

data
a tbl or data frame.

vars
a list of quoted variables.

copy
If TRUE, will make copy of input.

x
an object to check

Examples

library(dplyr, warn.conflicts = FALSE)
if (require("nycflights13")) {
flights_dt <- tbl_dt(flights)
group_size(group_by(flights_dt, year, month, day))
group_size(group_by(flights_dt, dest))

monthly <- group_by(flights_dt, month)
summarise(monthly, n = n(), delay = mean(arr_delay))
}

#> Loading required package: nycflights13

#> Source: local data table [12 x 3]
#> 
#> # A tibble: 12 x 3
#>    month     n delay
#>    <int> <int> <dbl>
#>  1     1 27004    NA
#>  2    10 28889    NA
#>  3    11 27268    NA
#>  4    12 28135    NA
#>  5     2 24951    NA
#>  6     3 28834    NA
#>  7     4 28330    NA
#>  8     5 28796    NA
#>  9     6 28243    NA
#> 10     7 29425    NA
#> 11     8 29327    NA
#> 12     9 27574    NA