Join data table tbls.

See join for a description of the general purpose of the functions.

inner_join.data.table(x, y, ..., by = NULL, copy = FALSE,
  suffix = c(".x", ".y"))

left_join.data.table(x, y, ..., by = NULL, copy = FALSE,
  suffix = c(".x", ".y"))

right_join.data.table(x, y, ..., by = NULL, copy = FALSE,
  suffix = c(".x", ".y"))

full_join.data.table(x, y, ..., by = NULL, copy = FALSE,
  suffix = c(".x", ".y"))

semi_join.data.table(x, y, by = NULL, copy = FALSE, ...)

anti_join.data.table(x, y, by = NULL, copy = FALSE, ...)

Arguments

x, y
tbls to join

...
For `inner_join()`, `left_join()`, `right_join()` and `full_join()` passed on to data.table [merge()] method. For `semi_join()` and `anti_join()` included only for compatibility with generic and must be empty.

by
a character vector of variables to join by. If NULL, the default, *_join() will do a natural join, using all variables with common names across the two tables. A message lists the variables so that you can check they're right (to suppress the message, simply explicitly list the variables that you want to join).

To join by different variables on x and y use a named vector. For example, by = c("a" = "b") will match x.a to y.b.

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.

Examples

library(dplyr, warn.conflicts = FALSE)

if (require("Lahman")) {
batting_dt <- tbl_dt(Batting)
person_dt <- tbl_dt(Master)

# Inner join: match batting and person data
inner_join(batting_dt, person_dt)

# Left join: keep batting data even if person missing
left_join(batting_dt, person_dt)

# Semi-join: find batting data for top 4 teams, 2010:2012
grid <- expand.grid(
  teamID = c("WAS", "ATL", "PHI", "NYA"),
  yearID = 2010:2012)
top4 <- semi_join(batting_dt, grid, copy = TRUE)

# Anti-join: find batting data with out player data
anti_join(batting_dt, person_dt)
}

#> Loading required package: Lahman

#> Joining, by = "playerID"

#> Joining, by = "playerID"

#> Joining, by = c("yearID", "teamID")

#> Joining, by = "playerID"

#> Source: local data table [0 x 22]
#> 
#> # A tibble: 0 x 22
#> # … with 22 variables: playerID <chr>, yearID <int>, stint <int>, teamID <fct>,
#> #   lgID <fct>, G <int>, AB <int>, R <int>, H <int>, X2B <int>, X3B <int>,
#> #   HR <int>, RBI <int>, SB <int>, CS <int>, BB <int>, SO <int>, IBB <int>,
#> #   HBP <int>, SH <int>, SF <int>, GIDP <int>