Group by with ordering

Uses [data.table::setkey()]; this guarantees that the output is sorted and typically increases performance when peforming multiple operations on the same grouping.

key_by(.data, ...)

Arguments

.data
A [lazy_dt]

...
Variables to key by

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(data.table)

#> 
#> Attaching package: ‘data.table’

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     between, first, last

mtcars2 <- lazy_dt(mtcars)

# group_by() doesn't order by group
mtcars2 %>% group_by(cyl) %>% summarise(mpg = mean(mpg))

#> Source: local data table [?? x 2]
#> Call:   `_DT1`[, .(mpg = mean(mpg)), by = .(cyl)]
#> 
#>     cyl   mpg
#>   <dbl> <dbl>
#> 1     6  19.7
#> 2     4  26.7
#> 3     8  15.1
#> 
#> # Use as.data.table()/as.data.frame()/as_tibble() to access results

# key_by() does
mtcars2 %>% key_by(cyl) %>% summarise(mpg = mean(mpg))

#> Source: local data table [?? x 2]
#> Call:   setkeyv(copy(`_DT1`), cols = "cyl")[, .(mpg = mean(mpg)), by = .(cyl)]
#> 
#>     cyl   mpg
#>   <dbl> <dbl>
#> 1     4  26.7
#> 2     6  19.7
#> 3     8  15.1
#> 
#> # Use as.data.table()/as.data.frame()/as_tibble() to access results