count.dtplyr_step {dtplyr}	R Documentation

Count observations by group

Description

This is a method for the dplyr count() generic. It is translated using .N in the j argument, and supplying groups to keyby as appropriate.

Usage

## S3 method for class 'dtplyr_step'
count(.data, ..., wt = NULL, sort = FALSE, name = NULL)

Arguments

.data
A lazy_dt()

...
<data-masking> Variables to group by.

wt
<data-masking> Frequency weights. Can be NULL or a variable:

If NULL (the default), counts the number of rows in each group.

If a variable, computes sum(wt) for each group.

sort
If TRUE, will show the largest groups at the top.

name
The name of the new column in the output.

If omitted, it will default to n. If there's already a column called n, it will error, and require you to specify the name.

Examples

library(dplyr, warn.conflicts = FALSE)

dt <- lazy_dt(dplyr::starwars)
dt %>% count(species)
dt %>% count(species, sort = TRUE)
dt %>% count(species, wt = mass, sort = TRUE)

[Package dtplyr version 1.2.1 Index]