filter.dtplyr_step {dtplyr}	R Documentation

Subset rows using column values

Description

This is a method for the dplyr arrange() generic. It is translated to the i argument of [.data.table

Usage

## S3 method for class 'dtplyr_step'
filter(.data, ..., .preserve = FALSE)

Arguments

.data
A lazy_dt().

...
<data-masking> Expressions that return a logical value, and are defined in terms of the variables in .data. If multiple expressions are included, they are combined with the & operator. Only rows for which all conditions evaluate to TRUE are kept.

.preserve
Ignored

Examples

library(dplyr, warn.conflicts = FALSE)

dt <- lazy_dt(mtcars)
dt %>% filter(cyl == 4)
dt %>% filter(vs, am)

dt %>%
  group_by(cyl) %>%
  filter(mpg > mean(mpg))

[Package dtplyr version 1.2.1 Index]