Create a "lazy" data.table

A lazy data.table lazy captures the intent of dplyr verbs, only actually performing computation when requested (with [connect()], [as.data.frame()], [data.table::as.data.table()], or [tibble::as_tibble()]). This allows dtplyr to convert dplyr verbs into as few data.table expressions as possible, which leads to a high performance translation.

lazy_dt(x, name = NULL)

Arguments

x
A data table (or something can can be coerced to a data table)

name
Optionally, supply a name to be used in generated expressions. For expert use only.

Examples

library(dplyr)

mtcars2 <- lazy_dt(mtcars)
mtcars2

#> Source: local data table [?? x 11]
#> Call:   `_DT1`
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results

mtcars2 %>% select(mpg:cyl)

#> Source: local data table [?? x 2]
#> Call:   `_DT1`[, .(mpg, cyl)]
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results

mtcars2 %>% select(x = mpg, y = cyl)

#> Source: local data table [?? x 2]
#> Call:   `_DT1`[, .(x = mpg, y = cyl)]
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results

mtcars2 %>% filter(x == 1) %>% select(mpg)

#> Source: local data table [?? x 1]
#> Call:   `_DT1`[x == 1, .(mpg)]
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results

mtcars2 %>% select(mpg) %>% filter(x == 1)

#> Source: local data table [?? x 1]
#> Call:   `_DT1`[, .(mpg)][x == 1, ]
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results

mtcars2 %>% mutate(x2 = x * 2, x4 = x2 * 2)

#> Source: local data table [?? x 13]
#> Call:   copy(`_DT1`)[, `:=`(x2 = x * 2)][, `:=`(x4 = x2 * 2)]
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results

mtcars2 %>% transmute(x2 = x * 2, x4 = x2 * 2)

#> Source: local data table [?? x 12]
#> Call:   copy(`_DT1`)[, `:=`(x2 = x * 2)][, .(x2, x4 = x2 * 2)]
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results

mtcars2 %>% filter(x == 1) %>% mutate(x2 = x * 2)

#> Source: local data table [?? x 12]
#> Call:   `_DT1`[x == 1, ][, `:=`(x2 = x * 2)]
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results

by_cyl <- mtcars2 %>% group_by(cyl)
by_cyl %>% summarise(x = mean(x))

#> Source: local data table [?? x 2]
#> Call:   `_DT1`[, .(x = mean(x)), by = .(cyl)]
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results

by_cyl %>% group_by(mpg) %>% mutate(x = mean(x))

#> Source: local data table [?? x 12]
#> Call:   `_DT1`[, `:=`(x = mean(x)), by = .(mpg)]
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results

by_cyl %>% filter(mpg < mean(mpg)) %>% summarise(hp = mean(hp))

#> Source: local data table [?? x 2]
#> Call:   `_DT1`[, .SD[mpg < mean(mpg), .(hp = mean(hp))], by = .(cyl)]
#> 
#> # Use as.data.table()/as.data.frame()/as.tibble() to access results