mutate.dtplyr_step {dtplyr}	R Documentation

Create and modify columns

Description

This is a method for the dplyr mutate() generic. It is translated to the j argument of [.data.table, using := to modify "in place". If .before or .after is provided, the new columns are relocated with a call to data.table::setcolorder().

Usage

## S3 method for class 'dtplyr_step'
mutate(.data, ..., .before = NULL, .after = NULL)

Arguments

.data
A lazy_dt().

...
<data-masking> Name-value pairs. The name gives the name of the column in the output, and the value should evaluate to a vector.

.before, .after
<tidy-select> Optionally, control where new columns should appear (the default is to add to the right hand side). See relocate() for more details.

Examples

library(dplyr, warn.conflicts = FALSE)

dt <- lazy_dt(data.frame(x = 1:5, y = 5:1))
dt %>%
  mutate(a = (x + y) / 2, b = sqrt(x^2 + y^2))

# It uses a more sophisticated translation when newly created variables
# are used in the same expression
dt %>%
  mutate(x1 = x + 1, x2 = x1 + 1)

[Package dtplyr version 1.2.1 Index]