summarise.dtplyr_step {dtplyr}	R Documentation

Summarise each group to one row

Description

This is a method for the dplyr summarise() generic. It is translated to the j argument of [.data.table.

Usage

## S3 method for class 'dtplyr_step'
summarise(.data, ..., .groups = NULL)

Arguments

.data
A lazy_dt().

...
<data-masking> Name-value pairs of summary functions. The name will be the name of the variable in the result.

The value can be:

A vector of length 1, e.g. min(x), n(), or sum(is.na(y)).

A vector of length n, e.g. quantile().

A data frame, to add multiple columns from a single expression.

.groups
Grouping structure of the result.

"drop_last": dropping the last level of grouping. This was the only supported option before version 1.0.0.

"drop": All levels of grouping are dropped.

"keep": Same grouping structure as .data.

When .groups is not specified, it defaults to "drop_last".

In addition, a message informs you of that choice, unless the result is ungrouped, the option "dplyr.summarise.inform" is set to FALSE, or when summarise() is called from a function in a package.

Examples

library(dplyr, warn.conflicts = FALSE)

dt <- lazy_dt(mtcars)

dt %>%
  group_by(cyl) %>%
  summarise(vs = mean(vs))

dt %>%
  group_by(cyl) %>%
  summarise(across(disp:wt, mean))

[Package dtplyr version 1.2.1 Index]