transmute.dtplyr_step {dtplyr}	R Documentation

Create new columns, dropping old

Description

This is a method for the dplyr transmute() generic. It is translated to the j argument of [.data.table.

Usage

## S3 method for class 'dtplyr_step'
transmute(.data, ...)

Arguments

.data
A lazy_dt().

...
<data-masking> Name-value pairs. The name gives the name of the column in the output, and the value should evaluate to a vector.

Examples

library(dplyr, warn.conflicts = FALSE)

dt <- lazy_dt(dplyr::starwars)
dt %>% transmute(name, sh = paste0(species, "/", homeworld))

[Package dtplyr version 1.2.1 Index]